requirements.txt
dm-haiku
blackjax-nightly
distrax
chex
matplotlib
optax
tqdm
pandas
jupyterlab
wandb
pytest
hydra-core
hydra-joblib-launcher
joblib
mdtraj
tables
git+https://github.com/VincentStimper/boltzmann-generators